requirements.txt
click==7.1.2
install==1.3.4
numpy==1.19.1
opencv-contrib-python==3.4.8.29
opencv-python==3.4.8.29
Pillow==8.3.2
six==1.15.0
toml==0.10.1
tabulate==0.8.7
prettytable==1.0.1
face-recognition==1.3.0